Fixes #17393 - The 'start' key always needs a 1 or 0
Previously, the docs mentioned that the 'start' key for compute-attributes is
a boolean. This suggests values like 'true', 'false', 'yes' and 'no' would be
accepted.

This was not the case. For libvirt (which is what I tested with, I assume oVirt
is the same), a 'true' or 'false' value did nothing, only a 1 or 0 did.

This patch fixes the docs to explain a 0 or 1 is required to influence
autostart behaviour.