Fixes #18739 - Display hostgroup title on host actions
Before this patch, host list was displaying only the leaf of the hostgroup tree.
E.g: 'base/hostgroup1' displayed as 'hostgroup1' in the Host list table.

After this patch, 'base/hostgroup1' is displayed, which is more useful
than the leaf (which could be the same for different trees). The same
thing has been fixed for Host info.