Fixes #21768 - Allow nil searchables
Adds ability to correctly resolve name params set to NIL.

For better coexistence with hammer defaults the logic
was moved to option sources which is now preferred place
for manipulation with options. As a result when options
or all_options is called anywhere in the code it should
return fully resolved options in final state.