Fixes #28420: Unpin mocha version & fix deprecation warning
The regression introduced in mocha v1.10.0 which led to its version
being pinned to v1.9.0 was fixed in mocha v1.11.2 which is currently the
latest version, so it is nowsafe to unpin the version.

Requiring 'mocha/setup' was deprecated in v1.10.0, so I've also fixed
that with the change to `test/test_helper.rb`.